Extract mutations, middleware, and geometry utils
Refactor AdminDashboard and backend to improve code organization:

- Extract people/member/volunteer mutations into usePeopleMutations hook
- Extract registration mutations into useRegistrationAdminMutations hook
- Extract venue/layout/table mutations into useVenueMutations hook
- Extract CORS middleware into app/middleware.py with focused test
- Extract geometry utilities into layoutGeometry.ts with unit tests
- Reduce AdminDashboard from 985 lines to ~240 by delegating mutations
- Update OPTIMIZATION-PLAN.md to mark steps 1-8 as complete

All new hooks and utilities include comprehensive tests. No behavioral changes.
